+++ src/kspdg/__init__.py
@@ -5,7 +5,7 @@
 # Single-sourcing package version
 # https://packaging.python.org/guides/single-sourcing-package-version/
 
-__version__ = "0.9.0-alpha"
+__version__ = "0.9.0-beta"
 
 # these imports make the individual environments accessible at the top-level
 # of the library and assign an environment version number
@@ -41,78 +41,38 @@ from kspdg.sb1.e1_envs import SB1_E1_I2_Env as SB1_E1_I2_V1
 from kspdg.sb1.e1_envs import SB1_E1_I3_Env as SB1_E1_I3_V1
 from kspdg.sb1.e1_envs import SB1_E1_I4_Env as SB1_E1_I4_V1
 from kspdg.sb1.e1_envs import SB1_E1_I5_Env as SB1_E1_I5_V1
-
-# Private-source, python-version-specific, platform and architecture-specific
-# environments with advanced bots (e.g. julia-based)
-import sys
-import platform
-import importlib
-def get_python_version():
-    # Get the Python version in the format 'python3_12'
-    version_info = sys.version_info
-    return f"python{version_info.major}_{version_info.minor}"
-
-def get_platform_architecture():
-    # Get the platform and architecture information
-    system_platform = platform.system()
-    machine = platform.machine()
-
-    if system_platform == 'Darwin':
-        system_platform = 'Darwin'
-    elif system_platform == 'Linux':
-        system_platform = 'Linux'
-    elif system_platform == 'Windows':
-        system_platform = 'Windows'
-    else:
-        raise RuntimeError(f"Unsupported platform: {system_platform}")
-
-    # Handle architecture
-    if machine == 'x86_64':
-        architecture = 'x86_64'
-    elif machine in ['arm64', 'armv8']:
-        architecture = 'arm64'
-    else:
-        raise RuntimeError(f"Unsupported architecture: {machine}")
-
-    return f"{system_platform}_{architecture}"
-
-def get_private_src_module_str(mod_name):
-    """returns string of module location to obfuscated code based on python version and system architecture"""
-    # Get dynamic parts of the import path
-    python_version = get_python_version()
-    platform_architecture = get_platform_architecture()
-
-    # Build the module path
-    return f"kspdg.private_src.{python_version}.{platform_architecture}.{mod_name}"
-
     
-# import obfuscated LBG1-LG3 environments
-__lg3_envs_path = get_private_src_module_str("kspdg_envs.lbg1.lg3_envs")
-try:
-    __lg3_envs_module = importlib.import_module(__lg3_envs_path)
-except ModuleNotFoundError:
-    print(f"Module {__lg3_envs_path} not found.")
-LBG1_LG3_I1_V1 = getattr(__lg3_envs_module, 'LBG1_LG3_I1_Env')
-LBG1_LG3_I2_V1 = getattr(__lg3_envs_module, 'LBG1_LG3_I2_Env')
+
+# import functions but name-mangle them so they are not 
+# inadvertently top-level accessible in the kspdg package
+import importlib
+from importlib.util import find_spec
+from kspdg.utils.private_src_utils import get_private_src_module_str as __get_mod_str
+
 
 # import obfuscated evaluate.py
-__evaluate_path = get_private_src_module_str("kspdg_envs.dist_evaluate")
+__evaluate_path = __get_mod_str("kspdg_envs.dist_evaluate")
 try:
     evaluate = importlib.import_module(__evaluate_path)
 except ModuleNotFoundError:
     print(f"Module {__evaluate_path} not found.")
-    
-# current_platform = platform.system()
-# architecture = platform.machine()
-# platform_architecture = f"{current_platform}_{architecture}"
-# if sys.version_info[:2] == (3, 12):
-#     # Python 3.12a
-#     from kspdg.private_src.python3_12.kspdg_envs.lbg1.lg3_envs import LBG1_LG3_I1_Env as LBG1_LG3_I1_V1
-#     from kspdg.private_src.python3_12.kspdg_envs.lbg1.lg3_envs import LBG1_LG3_I2_Env as LBG1_LG3_I2_V1
-# elif sys.version_info[:2] == (3, 9):
-#     # Python 3.9
-#     from kspdg.private_src.python3_9.kspdg_envs.lbg1.lg3_envs import LBG1_LG3_I1_Env as LBG1_LG3_I1_V1
-#     from kspdg.private_src.python3_9.kspdg_envs.lbg1.lg3_envs import LBG1_LG3_I2_Env as LBG1_LG3_I2_V1
-# else:
-#     # Handle other versions or raise an error
-#     raise ImportError(f"Private-source environments require python 3.9 or 3.12, got {sys.version}")
+
+# Condition import of LG3 environments on the presence of juliacall dependency
+# Therefore, the kspdg library should be usable without the adv_bots optional 
+# dependency
+if find_spec('juliacall') is not None:
+    # import obfuscated LBG1-LG3 environments
+    __lg3_envs_path = __get_mod_str("kspdg_envs.lbg1.lg3_envs")
+    try:
+        __lg3_envs_module = importlib.import_module(__lg3_envs_path)
+    except ModuleNotFoundError:
+        print(f"Module {__lg3_envs_path} not found.")
+    LBG1_LG3_I1_V1 = getattr(__lg3_envs_module, 'LBG1_LG3_I1_Env')
+    LBG1_LG3_I2_V1 = getattr(__lg3_envs_module, 'LBG1_LG3_I2_Env')
+
+else:
+    LBG1_LG3_I1_V1 = LBG1_LG3_I2_V1 = lambda *args, **kwargs: (
+        "Unmet dependency juliacall for using LBG1_LG3 environments.\n" +
+        "Please install kspdg[adv_bots] or kspdg[full] to use these.\n" +
+        "Refere to README for further instructions."
+    )